registration & login

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks,    
-        ]);
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,    
+            ];
+            
+            return view('tasks.index', $data);
+        } else {
+            return view('welcome');   
+        }
     }
 
     /**
@@ -29,11 +37,12 @@ class TasksController extends Controller
      */
     public function create()
     {
-        $task = new Task;
+        if (\Auth::check()) {
+            return view('tasks.create');
+        } else {
+            return redirect('/');
+        }
         
-        return view('tasks.create', [
-            'task' => $task,    
-        ]);
     }
 
     /**
@@ -44,16 +53,18 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'content' => 'required|max:191',
-            'status' => 'required|max:10',
-        ]);
+        if (\Auth::check()) {
+            $this->validate($request, [
+                'content' => 'required|max:191',
+                'status' => 'required|max:10',
+            ]);
+            
+            $request->user()->tasks()->create([
+                'content' => $request->content,
+                'status' => $request->status,            
+            ]);            
+        }
 
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
         return redirect('/');
     }
 
@@ -67,9 +78,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.show', [
-           'task' => $task, 
-        ]);
+        if ($task->user_id === \Auth::id()) {    
+            return view('tasks.show', [
+               'task' => $task, 
+            ]);            
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -81,10 +96,14 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
-        
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+            
+        if ($task->user_id === \Auth::id()) {    
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);            
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -96,16 +115,21 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'content' => 'required|max:191',
-            'status' => 'required|max:10',
-        ]);
+        if (\Auth::check()) {
+            $this->validate($request, [
+                'content' => 'required|max:191',
+                'status' => 'required|max:10',
+            ]);
+    
+            $task = Task::find($id);
+            
+            if ($task->user_id === \Auth::id()) {
+                $task->content = $request->content;
+                $task->status = $request->status;
+                $task->save();                            
+            }
+        }
 
-        $task = Task::find($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
         return redirect('/');
     }
 
@@ -118,8 +142,11 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
-        
+            
+        if ($task->user_id === \Auth::id()) {
+            $task->delete();            
+        }
+
         return redirect('/');
     }
 }
